EarlyStoppingTrainer bug: convert y to float

+++ src/maltorch/trainers/early_stopping_pytorch_trainer.py
@@ -96,7 +96,7 @@ class EarlyStoppingPyTorchTrainer:
             self._optimizer.zero_grad()
             outputs = model(x)
             outputs = outputs.squeeze()
-            loss = self._loss(outputs, y)
+            loss = self._loss(outputs, y.float())
             loss.backward()
             self._optimizer.step()
 
@@ -139,7 +139,7 @@ class EarlyStoppingPyTorchTrainer:
                 x, y = x.to(device), y.to(device)
                 outputs = model(x)
                 outputs = outputs.squeeze()
-                loss = self._loss(outputs, y)
+                loss = self._loss(outputs, y.float())
                 running_loss += loss.item()
                 y_preds = outputs.round()
